# docs/swagger.yaml
basePath: /api
definitions:
...
  ds.BookCreateRequest:
    properties:
      avg_sentence_len:
        type: number
      avg_word_len:
        type: number
      conjunction_freq:
        type: number
      image:
        type: string
      lexical_diversity:
        type: number
      text:
        type: string
      title:
        type: string
    required:
    - text
    - title
    type: object
  ds.BookDTO:
    properties:
      avg_sentence_len:
        type: number
      avg_word_len:
        type: number
      conjunction_freq:
        type: number
      id:
        type: integer
      image:
        type: string
      lexical_diversity:
        type: number
      status:
        type: boolean
      text:
        type: string
      title:
        type: string
    type: object
...
  ds.BookUpdateRequest:
    properties:
      avg_sentence_len:
        type: number
      avg_word_len:
        type: number
      conjunction_freq:
        type: number
      image:
        type: string
      lexical_diversity:
        type: number
      status:
        type: boolean
      text:
        type: string
      title:
        type: string
    type: object
...
  ds.PaginatedResponse:
    properties:
      items: {}
      total:
        type: integer
    type: object
...
host: localhost:8090
info:
  contact:
    email: support@example.com
    name: API Support
  description: API-сервер для управления книгами, заявками и анализом текстов
  title: API для системы анализа книг
  version: "1.0"
paths:
...
  /analysebooks/draft/books/{book_id}:
    post:
      description: Находит или создает черновик заявки для текущего пользователя и
        добавляет в него книгу.
      parameters:
      - description: ID книги для добавления
        in: path
        name: book_id
        required: true
        type: integer
      responses:
        "201":
          description: Сообщение об успехе
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Добавить книгу в черновик заявки (все)
      tags:
      - books
...
  /books:
    get:
      description: Возвращает постраничный список книг.
      parameters:
      - description: Фильтр по названию книги
        in: query
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.PaginatedResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список книг (все)
      tags:
      - books
    post:
      consumes:
      - application/json
      description: Создает новую запись о книге.
      parameters:
      - description: Данные новой книги
        in: body
        name: bookData
        required: true
        schema:
          $ref: '#/definitions/ds.BookCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ds.BookDTO'
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен (не модератор)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Создать новую книгу (только модератор)
      tags:
      - books
  /books/{id}:
    delete:
      description: Удаляет книгу из системы.
      parameters:
      - description: ID книги для удаления
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить книгу (только модератор)
      tags:
      - books
    get:
      description: Возвращает детальную информацию о книге.
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.BookDTO'
        "404":
          description: Книга не найдена
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить одну книгу по ID (все)
      tags:
      - books
    put:
      consumes:
      - application/json
      description: Обновляет информацию о существующей книге.
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/ds.BookUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.BookDTO'
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Обновить книгу (только модератор)
      tags:
      - books
  /books/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Загружает и привязывает изображение к книге.
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: integer
      - description: Файл изображения
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: URL загруженного изображения
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Файл не предоставлен
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Необходима авторизация
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Загрузить изображение для книги (только модератор)
      tags:
      - books
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"